+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## v1.1.0
 - Adds option `onLoad` in `useForm()`
-- Fixes infinite loop with initialValues in `useForm()`
-- Fixes display of children in `<Field />` when the attribute `options` is undefined and using a custom component
+- Fixes infinite loop with `initialValues` in `useForm()`
+- Fixes display of children in `<Field>` when the attribute `options` is undefined and using a custom component
+- Fixes React warnings of duplicate keys with `options` in `<Field>`
 - Updates dependencies
 
 ## v1.0.0